move http module to its own file
